+++ src/client.ts
@@ -8,18 +8,17 @@ import { type Retry, type RetryConfig, retry } from "@infra-blocks/retry";
 import { trusted } from "@infra-blocks/types";
 import type { CommandOutput } from "./commands/command/command.js";
 import type { DynamoDbClientCommand } from "./commands/command/commands.js";
-import { DeleteItem } from "./commands/command/delete-item.js";
-import { GetItem } from "./commands/command/get-item.js";
-import type { DeleteItemInput } from "./commands/command/inputs/delete-item.js";
-import type { DeleteItemOutput } from "./commands/command/outputs/delete-item.js";
-import {
-  DeleteTable,
-  type DeleteTableParams,
-} from "./commands/delete.table.js";
 import {
   CreateTable,
   type CreateTableInput,
   type CreateTableOutput,
+  DeleteItem,
+  type DeleteItemInput,
+  type DeleteItemOutput,
+  DeleteTable,
+  type DeleteTableInput,
+  type DeleteTableOutput,
+  GetItem,
   type GetItemInput,
   type GetItemOutput,
 } from "./commands/index.js";
@@ -132,17 +131,16 @@ export class DynamoDbClient {
   /**
    * Deletes a table using the DeleteTable API.
    *
-   * @param params - The parameters to use to delete the table.
+   * @param input - The parameters to use to delete the table.
    *
    * @see https://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_DeleteTable.html
    */
-  async deleteTable(params: DeleteTableParams): Promise<void> {
+  async deleteTable(input: DeleteTableInput): Promise<DeleteTableOutput> {
     if (this.logger.isDebugEnabled()) {
-      this.logger.debug("deleteTable(%s)", JSON.stringify(params));
+      this.logger.debug("deleteTable(%s)", JSON.stringify(input));
     }
 
-    const command = DeleteTable.from(params);
-    await this.client.send(command.toAwsCommand());
+    return this.send(new DeleteTable(input));
   }
 
   /**