+++ src/commands/command/index.ts
@@ -1,6 +1,7 @@
 export type { CommandInput, CommandOutput } from "./command.js";
 export { CreateTable } from "./create-table.js";
 export { DeleteItem } from "./delete-item.js";
+export { DeleteTable } from "./delete-table.js";
 export { GetItem } from "./get-item.js";
 export * from "./inputs/index.js";
 export * from "./outputs/index.js";