+++ src/commands/command/inputs/delete-table.ts
@@ -1,0 +1,11 @@
+import type { DeleteTableCommandInput } from "@aws-sdk/client-dynamodb";
+
+export interface DeleteTableInput {
+  name: string;
+}
+
+export function encode(input: DeleteTableInput): DeleteTableCommandInput {
+  return {
+    TableName: input.name,
+  };
+}