+++ src/commands/command/outputs/index.ts
@@ -1,3 +1,4 @@
 export type { CreateTableOutput } from "./create-table.js";
 export type { DeleteItemOutput } from "./delete-item.js";
+export type { DeleteTableOutput } from "./delete-table.js";
 export type { GetItemOutput } from "./get-item.js";