+++ test/integration/commands/delete-table.spec.ts
@@ -1,9 +1,5 @@
 import { DynamoDBClient, ListTablesCommand } from "@aws-sdk/client-dynamodb";
 import { expect } from "@infra-blocks/test";
-import type {
-  CreateTableInput,
-  DeleteTableParams,
-} from "../../../src/index.js";
 import { dropAllTables } from "../fixtures.js";
 
 describe(DynamoDBClient.name, () => {
@@ -12,17 +8,17 @@ describe(DynamoDBClient.name, () => {
   describe("deleteTable", () => {
     it("should fail for non-existent table", async function () {
       const client = this.createClient();
-      const params: DeleteTableParams = {
+      const params = {
         name: "test-table",
       };
       await expect(client.deleteTable(params)).to.be.rejected;
     });
     it("should work with existing table", async function () {
       const client = this.createClient();
-      const params: CreateTableInput = {
+      const params = {
         name: "test-table",
         keySchema: { partitionKey: { name: "pk", type: "S" } },
-      };
+      } as const;
       await client.createTable(params);
       await client.deleteTable({ name: params.name });